+++ lib/alpaca/trade/api/clock.rb
@@ -4,11 +4,11 @@ module Alpaca
   module Trade
     module Api
       class Clock
-        attr_reader :timestamp, :open, :next_open, :next_close
+        attr_reader :timestamp, :is_open, :next_open, :next_close
 
         def initialize(json)
           @timestamp = json['timestamp']
-          @open = json['open']
+          @is_open = json['is_open']
           @next_open = json['next_open']
           @next_close = json['next_close']
         end